requirements.txt
sounddevice
soundfile
numpy
torch
openai-whisper
ollama
colorama
Flask
PyYAML
TTS